Module mercator_db::space

Reference space definitions.

This include notions such as shapes, positions, axes, etc…

Structs

Axis
Definition of an axis of a base.

Graduation
Definition of a fixed-precision, finite length axis.

Space
A reference space, defined by its name and coordinate system.

Enums

Coordinate
Store efficiently a coordinate.

CoordinateSystem
Kinds of space coordinate systems, or bases

NumberSet
Mathematical set numbers.

Position
Store a position as efficiently as possible in terms of space.

Shape
Known shapes descriptions