Enum mercator_db::space::Position

pub enum Position {
    Position1(Coordinate),
    Position2([Coordinate; 2]),
    Position3([Coordinate; 3]),
    Position4([Coordinate; 4]),
    Position5([Coordinate; 5]),
    Position6([Coordinate; 6]),
    Position7([Coordinate; 7]),
    Position8([Coordinate; 8]),
    PositionN(Vec<Coordinate>),
}

Store a position as efficiently as possible in terms of space.

Variants

Position1(Coordinate)

1 dimension positions.

Position2([Coordinate; 2])

2 dimensions positions.

Position3([Coordinate; 3])

3 dimensions positions.

Position4([Coordinate; 4])

4 dimensions positions.

Position5([Coordinate; 5])

5 dimensions positions.

Position6([Coordinate; 6])

6 dimensions positions.

Position7([Coordinate; 7])

7 dimensions positions.

Position8([Coordinate; 8])

8 dimensions positions.

PositionN(Vec<Coordinate>)

N dimensions positions.

Methods

impl Position

pub fn dimensions(&self) -> usize

Returns the number of dimensions or size of the vector.

pub fn norm(&self) -> f64

Compute ||self||.

pub fn unit(&self) -> Self

Compute the unit vector pointing in the same direction as self.

pub fn dot_product(&self, rhs: &Self) -> f64

Multiplies self with rhs, producing a scalar value.

self • rhs = product

Note: The two vector sizes must be equal, a.k.a the two vectors must have the same number of dimensions.

Parameters

rhs: The right-hand side vector.

pub fn reduce_precision(&self, scale: u32) -> Self

Remove bits of precision.

Parameters

scale: Number of bits of precision to remove from each coordinates.