Function mercator_db::storage::model::build_index

pub fn build_index(
    name: &str, 
    version: &str, 
    spaces: &[Space], 
    objects: &[SpatialObject], 
    scales: Option<Vec<Vec<u32>>>, 
    max_elements: Option<usize>
) -> Result<Core, String>

Generate an index.

Parameters

name: Name to give to the index.

version: Parameter to distinguish revisions of an index.

spaces: A list of the reference spaces. Only objects whose reference space is known will be indexed.

objects: The data points to index.

scales: An optional list of specific index resolutions to generates on top of the full resolution one.

max_elements: If this is specified, automatically generates scaled indices, by halving the number elements between resolutions, and stop generating indices either when the number of points remaining is equal to the number of distinct Ids, or smaller or equal to this value.

Note: max_elements is ignored when scales is not None.