Module mercator_db::storage::model::v2

REST API objects, v2.

Structs

SpatialObject
Links Properties to a list of spatial volumes.

Volume
Defines a volume as the union of geometric shapes.

Enums

Shape
Describes an homogeneous list of geometric shapes.

Functions

to_spatial_objects
Convert a list of properties grouped by space id, then positions to a list of Spatial Objects for the rest API v2.