Function mercator_db::storage::model::v2::to_spatial_objects

pub fn to_spatial_objects(
    list: Vec<(&String, Vec<(Position, &Properties)>)>
) -> Vec<SpatialObject>

Convert a list of properties grouped by space id, then positions to a list of Spatial Objects for the rest API v2.

Parameters

list: A list of (Space Id, [ ( Spatial position, &Properties ) ]) tuples.