Struct mercator_db::Core

pub struct Core { /* fields omitted */ }

Index over a single dataset

Methods

impl Core

pub fn new<S>(
title: S,
version: S,
spaces: &[Space],
properties: Vec<Properties>,
space_objects: Vec<SpaceSetObject>,
scales: Option<Vec<Vec<u32>>>,
max_elements: Option<usize>
) -> Result<Self, String> where
S: Into<String>,

Instantiate a new index for a dataset.

Parameters

title: The title to use for the new dataset.

version: The revision of the new dataset.

spaces: The list of reference spaces used within the dataset.

properties: The identifiers, has an ordered list, which is referenced by the space_objects by offset within this list.

space_objects: A list of links between volumetric positions and identifiers.

scales: A list of resolutions for which to build indices. Each value represent the number of bits of precision to remove from the coordinates to build the index.

max_elements: The minimum number of positions to use as a stopping condition while building automatically multiple resolutions of the index.

Each consecutive index will contains at most half the number of data points than the next finer-grained index.

The minimum number of elements contained within an index is this value or the number of identifiers, whichever is greater.

pub fn name(&self) -> &String

Title of the dataset.

pub fn version(&self) -> &String

Revision of the dataset.

pub fn keys(&self) -> &Vec<Properties>

List of identifiers contained in this dataset.

pub fn get_by_positions(
&self,
parameters: &CoreQueryParameters,
positions: &[Position],
space_id: &str
) -> ResultSet

Retrieve everything located at specific positions.

Parameters

parameters: Search parameters, see CoreQueryParameters.

positions: Volume to use to filter data points.

space_id: positions are defined as decoded coordinates in this reference space.

pub fn get_by_shape(
&self,
parameters: &CoreQueryParameters,
shape: &Shape,
space_id: &str
) -> ResultSet

Search using a shape which defines a volume.

Parameters

parameters: Search parameters, see CoreQueryParameters.

shape: Volume to use to filter data points.

space_id: shape is defined as decoded coordinates in this reference space.

pub fn get_by_id<S>(
&self,
parameters: &CoreQueryParameters,
id: S
) -> Result<Vec<(&String, Vec<Position>)>, String> where
S: Into<String>,

Search by Id, a.k.a retrieve all the positions linked to this id.

Parameters

parameters: Search parameters, see CoreQueryParameters.

id: Identifier for which to retrieve is positions.

pub fn get_by_label<S>(
&self,
parameters: &CoreQueryParameters,
id: S
) -> ResultSet where
S: Into<String>,

Search by label, a.k.a use an identifier to define the search volume.

Parameters

parameters: Search parameters, see CoreQueryParameters.

id: Identifier to use to define the search volume.